+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/dataset/hibernate/HibernateCompleteDataSetRegistrationStore.java
@@ -180,6 +180,7 @@ public class HibernateCompleteDataSetRegistrationStore
 
   @Override
   public void deleteByCategoryOptionCombo(@Nonnull Collection<CategoryOptionCombo> cocs) {
+    if (cocs.isEmpty()) return;
     String hql =
         """
         delete from CompleteDataSetRegistration cdsr
